==> Postwork/1_Analisis_poswork.ipynb <==
## Análisis Exploratorio de nuestro Dataset

### 1. ¿El conjunto de datos que tengo realmente me sirve para responder algunas de las preguntas que me planteé?

1. **"100_mil_programadores_2020_Mintic.json":**
    Se está leyendo el conjundo de datos publicado por el Mintic "100_mil_programadores_2020_Mintic.json", donde se encuentra información de la población estudiantil en el 2020 que se inscribió al progrma 100 mil programadores. Este programa se creo con el fin de suplir la demanda del mercado y el déficit de programadores en Colombia.

2. **"Saber-11-2020-1.json" y "Saber-11-2020-2.json":**
    Resultados anonimizados de las pruebas de Saber 11 de Calendario A y B del año 2020. Que pueden dar información de las compencias según la población en la cual se encuentran los estudiantes mejor preparados

### 2. ¿Qué tamaño tiene mi conjunto de datos? ¿Serán datos suficientes?
1. **"100_mil_programadores_2020_Mintic.json":**

import json
import pandas as pd
import pprint

# Paso del conjunto de datos "100 mil programadores 2020" a Data Frame
datos_json = json.load(open("100_mil_programadores_2020_Mintic.json", encoding="utf8"))
df_mintic = pd.DataFrame(datos_json["data"])

# El primer conjunto de datos contiene 429.699 registros. Cada registro corresponde a un estudiante
df_mintic.shape

2. **"Saber-11-2020-1.json" y "Saber-11-2020-2.json" :**

# Paso del conjunto de datos "Saber-11-2020-1.json" a Data Frame
datos_json_2 = json.load(open("saber_11_2020_1.json", encoding="utf8"))
df_saber_1 = pd.DataFrame(datos_json_2["data"])

# El primer conjunto de datos contiene 15.435 registros. Cada registro corresponde a un estudiante
df_saber_1.shape

# Paso del conjunto de datos "Saber-11-2020-1.json" a Data Frame
datos_json_3 = json.load(open("saber_11_2020_2.json", encoding="utf8"))
df_saber_2 = pd.DataFrame(datos_json_3["data"])

# El primer conjunto de datos contiene 504.872 registros. Cada registro corresponde a un estudiante
df_saber_2.shape

### 3. ¿Qué columnas tengo y qué información tengo en cada una de esas columnas?

Columnas e información de **100_mil_programadores_2020_Mintic.json**:

En el json se encuentra la información meta data con el nombre de las columnas, en total son 23 en las cuales se encuentra información como: 
* Ubicación del estudaiante
* Estrato social
* Edad
* Género
* Grado
* Ruta a la que se inscribió (1 o 2) según el programa

dic_meta = datos_json["meta"]["view"]
lista_columnas_mintic = [i["name"] for i in dic_meta["columns"]]
pprint.saferepr(lista_columnas_mintic)

Columnas e información de **Saber-11-2020-1.json" y "Saber-11-2020-2.json**:

Ambos conjuntos de información aunque tienen diferente cantidad de columnas coinciden en la mayoría. Estas columnas tienen información de los resultados pruebas saber 11 de los calendarios 1 y 2, entre ellos:
* Género del estudiante 'ESTU_GENERO'
* Fecha de nacimiento 'ESTU_FECHANACIMIENTO'
* Ubicación de residencia ESTU_PAIS_RESIDE', 'ESTU_DEPTO_RESIDE', 'ESTU_MCPIO_RESIDE'.
* Estrato 'FAMI_ESTRATOVIVIENDA'
* Condición de educación y económica de la familia
* Puntajes de la prueba en cada área del conocimiento evaluada y puntajes en percentiles

dic_meta_2 = datos_json_2["meta"]["view"]
lista_columnas_saber_1 = [i["name"] for i in dic_meta_2["columns"]]
pprint.saferepr(lista_columnas_saber_1)

dic_meta_3 = datos_json_3["meta"]["view"]
lista_columnas_saber_2 = [i["name"] for i in dic_meta_3["columns"]]
pprint.saferepr(lista_columnas_saber_2)

#### Aplicación de nombres de columnas a dataFrames para cada conjunto de datos
Debido a que el nombre de las columnas y los datos se encuentran separados en la lectura del json. Se procede a unir cada conjunto de datos con los nombres de cada columna antes de realizar el análisis de los datos de manera general

##### Union de columnas y data para inscripción de 100 mil programadores del mintic

df_mintic.columns = lista_columnas_mintic
df_mintic.head(3)

##### Union de columnas y data para saber 11 periodo 1

df_saber_1.columns = lista_columnas_saber_1
df_saber_1.head(3)

##### Union de columnas y data para saber 11 periodo 2

df_saber_2.columns = lista_columnas_saber_2
df_saber_2.head(3)

### 4. Analisis de datos
* ¿Qué tipos de datos tengo en cada columna? 
    Se realiza la consulta de los tipos de cada columna. 
    
* ¿Parecen ser el tipo correcto de datos? ¿O es un tipo de datos "incorrecto"?
    Según se observa el tipo de dato corresponde para cada DF con la información presentada

* Si selecciono algunas filas al azar y las observo, ¿estoy obteniendo los datos que debería? ¿o hay datos que parecen estar "sucios" o "incorrectos"?
    Según algunas filas consultadas, su correspondencia es correcta, según los nombres de la columna y datos esperadas, para cada DF.

df_mintic.dtypes

df_saber_1.dtypes

df_saber_2.dtypes
