==> Postwork/3_ConcatenacionDataFrames.ipynb <==
## Concatenación DataFrames

### 4. Union Prueba Saber 11 periodo 1 y periodo 2

De acuerdo con la limpieza realizada para estas dos tablas y la información analizada, se realizará la union de los dos dataframes de las pruebas Saber 11 para un mejor análisis posterior.

#### Verificar tamaños de cada dataframe resultante

print("Data Frame Mintic: ", df_mintic_v2.shape)
print("Saber11 periodo 1: ", df_saber_1_v2.shape)
print("Saber11 periodo 2: ", df_saber_2_v2.shape)

Se realiza la concatenación de los Data Frames Saber11 periodos 1 y 2 ya que cuentan con la misma información

df_saber_union = pd.concat( [df_saber_1_v2, df_saber_2_v2] )

Se comprueba los datos de inicio y final del data Frame resultante df_saber_union

df_saber_union.head(3)

df_saber_union.tail(3)

#### Reindexación de Data Frame resultante

Por conveniencia se dejará el índice como la cantidad de registros en el data frame resultante

df_saber_union_v1 = df_saber_union.reset_index(drop=True)
df_saber_union_v1
