==> Postwork/4_TFO_datos.ipynb <==
# Transformación, filtración y ordenamiento de datos

### 1. Revisión de tipos de datos de los dos dataFrames

#### DF Mintic

df_mintic_v2.dtypes

df_mintic_v2.head(3)

Para el DF de Mintic se modificarán los tipos de datos:
* Estrato_social: moddificación a int
* Edad: modificación a int
* Grado: modificación a int

diccionario_de_conversion_mintic = {
    "ESTRATO_SOCIAL":"int64",
    "EDAD":"int64",
    "GRADO":"int64"
}

df_mintic_v3 = df_mintic_v2
try:
    df_mintic_v3 = df_mintic_v2.astype(diccionario_de_conversion_mintic)
except ValueError as e:
    print(type(e))
    
df_mintic_v3.dtypes

df_mintic_v3["ESTRATO_SOCIAL"].unique()

Me genera error en el estrato social al tratar de convertir a entero, ya que tiene un valor "a".
Se procederá a modificar todos los valores a de esta columna al valor 0, el cuan indicará que no tiene estrato social asignado

# Se aplica el cambio en la versión 2
df_mintic_v2.loc[df_mintic_v2.ESTRATO_SOCIAL=="a",'ESTRATO_SOCIAL'] = 0
df_mintic_v2["ESTRATO_SOCIAL"].unique()

df_mintic_v3 = df_mintic_v2
try:
    df_mintic_v3 = df_mintic_v2.astype(diccionario_de_conversion_mintic)
except ValueError as e:
    print(type(e))
    
df_mintic_v3.dtypes

df_mintic_v3.head(3)

#### DF Saber 11

df_saber_union_v1.dtypes

df_saber_union_v1.head(3)

### Se revisan las columnas de Saber 11 según conveniencia y compativilidad con Mintic

La columna "FAMI_ESTRATOVIVIENDA" se transforma a numerica según el texto, para tener relación con la tabla Mintic: El estrato 'None' y 'Sin Estrato', serán transformados en 0

df_saber_union_v1["FAMI_ESTRATOVIVIENDA"].unique()

#Revisando data de Mintic
df_mintic_v3["ESTRATO_SOCIAL"].unique()

conver_estrato = {
    'Estrato 1': 1,
    'Estrato 2': 2,
    'Estrato 3': 3,
    'Estrato 4': 4,
    'Estrato 5': 5,
    'Estrato 6': 6,
    'Sin Estrato': 0,
    None: 0,
}

df_saber_union_v1["FAMI_ESTRATOVIVIENDA"] = df_saber_union_v1["FAMI_ESTRATOVIVIENDA"].map(conver_estrato)
df_saber_union_v1["FAMI_ESTRATOVIVIENDA"].unique()

Tansformamos la fecha de nacimiento 'ESTU_FECHANACIMIENTO' a la edad debido a que la data de Mintic se encuentra como edad y lo dejaremos como entero

from datetime import datetime
from dateutil.relativedelta import relativedelta
import numpy as np

type(datetime.fromisoformat(df_saber_union_v1["ESTU_FECHANACIMIENTO"][0]))

fecha_2020 = datetime.fromisoformat('2020-01-01') # tipo datetime.datetime

# Serie Datetime
ser_nac = pd.Series( datetime.fromisoformat(i) for i in df_saber_union_v1["ESTU_FECHANACIMIENTO"] )
ser_nac.head(2)

Restar fecha 2020 a la fecha de nacimiento para conocer la edad

edades = pd.Series( relativedelta(fecha_2020, ser_nac[m]).years for m in range( len(ser_nac) ) )
edades

df_saber_union_v1["ESTU_FECHANACIMIENTO"] = edades
df_saber_union_v1["ESTU_FECHANACIMIENTO"].head(3)

### Se convertiran todas las columnas que contienen 'Si' y 'No' del dataframe Saber11

* FAMI_TIENEINTERNET
* FAMI_TIENECOMPUTADOR

### Se convertiran todas las columnas numericas a enteros en datafreme Seber 11
* PUNT_LECTURA_CRITICA
* PERCENTIL_LECTURA_CRITICA        
* DESEMP_LECTURA_CRITICA           
* PUNT_MATEMATICAS                 
* PERCENTIL_MATEMATICAS            
* DESEMP_MATEMATICAS               
* PUNT_C_NATURALES                 
* PERCENTIL_C_NATURALES            
* DESEMP_C_NATURALES               
* PUNT_SOCIALES_CIUDADANAS         
* PERCENTIL_SOCIALES_CIUDADANAS    
* DESEMP_SOCIALES_CIUDADANAS       
* PUNT_INGLES                      
* PERCENTIL_INGLES                 
* PUNT_GLOBAL                      
* PERCENTIL_GLOBAL                 

conver_sn = {'No': 0, 'Si': 1, None: None}
conver_sn

df_saber_union_v1["FAMI_TIENEINTERNET"] = df_saber_union_v1["FAMI_TIENEINTERNET"].map(conver_sn)
df_saber_union_v1["FAMI_TIENEINTERNET"].unique()

df_saber_union_v1["FAMI_TIENEINTERNET"] = df_saber_union_v1["FAMI_TIENEINTERNET"].astype('Int64')
df_saber_union_v1["FAMI_TIENEINTERNET"].unique()

df_saber_union_v1["FAMI_TIENECOMPUTADOR"] = df_saber_union_v1["FAMI_TIENECOMPUTADOR"].map(conver_sn)
df_saber_union_v1["FAMI_TIENECOMPUTADOR"].unique()

df_saber_union_v1["FAMI_TIENECOMPUTADOR"] = df_saber_union_v1["FAMI_TIENECOMPUTADOR"].astype('Int64')
df_saber_union_v1["FAMI_TIENECOMPUTADOR"].unique()

#### Convertir a enteros las calificaciones y puntajes

df_saber_union_v1['PUNT_LECTURA_CRITICA'] = df_saber_union_v1['PUNT_LECTURA_CRITICA'].astype(int)
df_saber_union_v1['PERCENTIL_LECTURA_CRITICA'] = df_saber_union_v1['PERCENTIL_LECTURA_CRITICA'].astype(int)
df_saber_union_v1['DESEMP_LECTURA_CRITICA'] = df_saber_union_v1['DESEMP_LECTURA_CRITICA'].astype(int)
df_saber_union_v1['PUNT_MATEMATICAS'] = df_saber_union_v1['PUNT_MATEMATICAS'].astype(int)
df_saber_union_v1['PERCENTIL_MATEMATICAS'] = df_saber_union_v1['PERCENTIL_MATEMATICAS'].astype(int)
df_saber_union_v1['DESEMP_MATEMATICAS'] = df_saber_union_v1['DESEMP_MATEMATICAS'].astype(int)
df_saber_union_v1['PUNT_C_NATURALES'] = df_saber_union_v1['PUNT_C_NATURALES'].astype(int)
df_saber_union_v1['PERCENTIL_C_NATURALES'] = df_saber_union_v1['PERCENTIL_C_NATURALES'].astype(int)
df_saber_union_v1['DESEMP_C_NATURALES'] = df_saber_union_v1['DESEMP_C_NATURALES'].astype(int)
df_saber_union_v1['PUNT_SOCIALES_CIUDADANAS'] = df_saber_union_v1['PUNT_SOCIALES_CIUDADANAS'].astype(int)
df_saber_union_v1['PERCENTIL_SOCIALES_CIUDADANAS'] = df_saber_union_v1['PERCENTIL_SOCIALES_CIUDADANAS'].astype(int)
df_saber_union_v1['DESEMP_SOCIALES_CIUDADANAS'] = df_saber_union_v1['DESEMP_SOCIALES_CIUDADANAS'].astype(int)
df_saber_union_v1['PERCENTIL_INGLES'] = df_saber_union_v1['PERCENTIL_INGLES'].astype(int)
df_saber_union_v1['PUNT_GLOBAL'] = df_saber_union_v1['PUNT_GLOBAL'].astype(int)

df_saber_union_v1.dtypes

lista_ing = [int(i) if type(i) == str else np.nan for i in df_saber_union_v1["PUNT_INGLES"]]
ser_int_ing = pd.Series(lista_ing).astype('Int64')
#df_saber_union_v1["PUNT_INGLES"] df_saber_union_v1["PUNT_INGLES"].head()
ser_int_ing.head(3)

df_saber_union_v1["PUNT_INGLES"] = ser_int_ing
df_saber_union_v1["PUNT_INGLES"].head()

lista_per_global = [int(i) if type(i) == str else np.nan for i in df_saber_union_v1["PERCENTIL_GLOBAL"]]
ser_int_glob = pd.Series(lista_per_global).astype('Int64')
ser_int_glob.head(3)

df_saber_union_v1["PERCENTIL_GLOBAL"] = ser_int_glob
df_saber_union_v1["PERCENTIL_GLOBAL"].head()

df_saber_union_v1.dtypes

df_saber_union_v1['ESTU_FECHANACIMIENTO'].unique()

Función para convertir edades no lógicas a vacíos

def conv_edad(edad):
    if edad > 99 or edad <= 9:
        return np.nan
    else:
        return edad

lista_edades_nuevas = [conv_edad(i) for i in df_saber_union_v1['ESTU_FECHANACIMIENTO']]

df_saber_union_v1['ESTU_FECHANACIMIENTO'] = lista_edades_nuevas
df_saber_union_v1['ESTU_FECHANACIMIENTO'].head(3)

ser_int_Fnac = pd.Series(df_saber_union_v1['ESTU_FECHANACIMIENTO']).astype('Int64')
ser_int_Fnac

df_saber_union_v1['ESTU_FECHANACIMIENTO'] = ser_int_Fnac
df_saber_union_v1['ESTU_FECHANACIMIENTO'].unique()

#### Cambio de nombre de la columna, ya que representa la edad y ya no representa la fecha de nacimiento

df_saber_union_v2 = df_saber_union_v1.rename(columns={'ESTU_FECHANACIMIENTO':'ESTU_EDAD'})
df_saber_union_v2.dtypes

[df_saber_union_v2[col].unique() for col in df_saber_union_v2.columns]

#### De la información resultante y de acuerdo para el análisis se dejaran los resultados solo de Colombia para analisar el programa del gobierno

df_saber_union_v3 = df_saber_union_v2[ df_saber_union_v2['ESTU_NACIONALIDAD'] == 'COLOMBIA']

df_saber_union_v4 = df_saber_union_v3[ df_saber_union_v3['ESTU_EDAD'] != np.nan]
df_saber_union_v4.head(3)

df_saber_union_v4.shape

df_saber_union_v4.to_csv('saber_11_union.csv', index=False)

df_mintic_v3.to_csv('Mintic_datos_limpios.csv', index=False)
