==> Postwork/5_Preguntas.ipynb <==
# Análisis y resolución de preguntas

import matplotlib.pyplot as plt
import numpy as np
import pandas as pd

##### Se realiza la carga de los datos ya transformados y limpiados para analizar las preguntas iniciales

df_mintic_final = pd.read_csv('Mintic_datos_limpios.csv')
df_mintic_final.head(3)

df_saber_11 = pd.read_csv('Saber_11_union.csv')

#### ¿Qué población se ha inscrito en el programa misión 100 mil programadores del MINTic en el 2020?

Registros al año 2020

df_mintic_final.groupby(['FECHA_CORTE']).size()

mintic_2020 = df_mintic_final[df_mintic_final['FECHA_CORTE'] == '2020-12-31']
mintic_2020.shape

Según el análisis de información siguiente:

De un total 137.388
* 128.857 aspirantes se inscribieron para el programa del Mintic
* 5.855 fueron Matriculados al programa

mintic_2020.groupby(['FECHA_CORTE', 'ESTADO_FORMACION']).size()

etiquetas = mintic_2020['ESTADO_FORMACION'].unique()
etiquetas

#### ¿Cuántos estudiantes de estudios básicos se postularon para entrar a las rutas 1 y 2 de la misión 100 mil programadores del MINTic en el 2020?

De los estudiantes aspirantes en el 2020:
* Matriculados - admitidos al programa en la **ruta 1**: 1.650
* Matriculados - admitidos al programa en la **ruta 2**: 4.205

mintic_2020.groupby(['RUTA', 'ESTADO_FORMACION']).size()

mintic_2020_RUTA1 = mintic_2020[mintic_2020['RUTA']== 'RUTA 1']
mintic_2020_RUTA1.groupby(['RUTA','ESTADO_FORMACION']).size()

#### ¿Qué resultados se obtuvieron en saber 11 para el 2020?

df_saber_11.head(3)

##### Se consultan y eliminan regstros que en su puntaje global dio como resultado 0 ya que no tiene información de puntajes en ningún área evaluada

df_saber_11_result0 = df_saber_11[df_saber_11['PUNT_GLOBAL'] == 0]
print('puntaje global: ',df_saber_11_result0['PUNT_GLOBAL'].unique())
print('puntaje percentil global: ',df_saber_11_result0['PERCENTIL_GLOBAL'].unique())
print('puntaje desempeño ciencias naturales: ',df_saber_11_result0['DESEMP_C_NATURALES'].unique())
print('puntaje desempeño ciencias ciudadanas: ',df_saber_11_result0['DESEMP_SOCIALES_CIUDADANAS'].unique())
print('puntaje desempeño ingles: ',df_saber_11_result0['DESEMP_INGLES'].unique())

df_saber_11_v2 = df_saber_11[df_saber_11['PUNT_GLOBAL'] != 0]
df_saber_11_v2.shape

df_saber_11_v2.columns

##### Interpretación de resultados pruebas Saber 11

* El puntaje global va de **0 a 500** puntos sin decimales.
* El puntaje de cada prueba va de **0 a 100**puntos sin decimales.
* Los niveles de desempeño definidos son: Insuficiente, Mínimo, Satisfactorio y Avanzado. Para todas las áreas, excepto Inglés. Son 4 niveles de desempeño donde 1 es el nivel más bajo y 4 el más alto.
* Los niveles de desempeño definidos para Inglés son: A-, A1, A2, B1, B+.
* Los niveles de desempeño definen lo que sabe y sabe hacer el estudiante.
* Los percentiles indican cómo se encuentra el resultado del evaluado con respecto a los resultados de los estudiantes del país: Ya no vas a encontrar puestos sino percentiles, estos se definen como un puesto relativo.

df_saber_11_v2.groupby(['ESTU_NACIONALIDAD'])['PUNT_GLOBAL'].agg(['mean','max','min'])

df_saber_11_v2.groupby(['ESTU_GENERO'])['PUNT_GLOBAL'].agg(['mean','max','min'])

#### ¿Cómo les va en las pruebas de matemáticas y lógica a los estudiantes 2020?

Puntaje matemáticas

df_saber_11_v2.groupby(['ESTU_GENERO'])['PUNT_MATEMATICAS'].agg(['mean','max','min'])

Desempeño matemáticas

df_saber_11_v2.groupby(['ESTU_GENERO'])['DESEMP_MATEMATICAS'].agg(['mean','max','min'])

### Análisis según condiciones de vivienda del estudiante para las pruebas matemáticas

#### Estrato vivivenda

df_saber_11_v2.groupby(['FAMI_ESTRATOVIVIENDA'])['PUNT_GLOBAL'].agg(['mean','max','min'])

df_saber_11_v2.groupby(['FAMI_ESTRATOVIVIENDA'])['PUNT_MATEMATICAS'].agg(['mean','max','min'])

df_saber_11_v2.groupby(['FAMI_ESTRATOVIVIENDA'])['DESEMP_MATEMATICAS'].agg(['mean','max','min'])

#### Si la familia tiene o no tiene Internet

df_saber_11_v2.groupby(['FAMI_TIENEINTERNET'])['PUNT_MATEMATICAS'].agg(['mean','max','min'])

df_saber_11_v2.groupby(['FAMI_TIENEINTERNET'])['DESEMP_MATEMATICAS'].agg(['mean','max','min'])

#### Si la familia tiene o no tiene computador

df_saber_11_v2.groupby(['FAMI_TIENECOMPUTADOR'])['PUNT_MATEMATICAS'].agg(['mean','max','min'])

df_saber_11_v2.groupby(['FAMI_TIENECOMPUTADOR'])['DESEMP_MATEMATICAS'].agg(['mean','max','min'])
